# backend/doc/patient-openapi.yaml
openapi: 3.0.3
info:
  title: Patient API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local backend server
paths:
...
  /api/doctor/patients:
    post:
      tags: [Patients]
      operationId: createPatientCard
      summary: Create patient card
      description: Creates a patient card from the authenticated doctor context using the region provided in the request.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatientRequest'
      responses:
        '201':
          description: Patient card created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedPatientResponse'
        '400':
          description: Validation failed
        '401':
          description: Authentication required
        '403':
          description: Only a doctor can create a patient card
    get:
      tags: [Patients]
      operationId: listDoctorPatients
      summary: List patients available to the doctor
      description: >
        Returns a paginated list of patients available to the authenticated doctor.
        scope=own returns only the doctor's own region with visible names.
        scope=all returns patients from all regions with anonymized names
        (lastName, firstName, middleName are null), and supports regionId and diagnosis filters.
        Each patient has a computed status based on the latest examination date:
        GREEN for less than 3 months, YELLOW for 3-6 months, RED for more than 6 months.
        The default ordering is RED first, then YELLOW, then GREEN.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: scope
          required: false
          description: Listing mode. own = doctor's region with names, all = all regions with anonymized names.
          schema:
            type: string
            enum: [own, all]
            default: own
        - in: query
          name: page
          required: false
          description: Zero-based page number.
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: limit
          required: false
          description: Number of patients per page.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: regionId
          required: false
          description: Optional region filter. Primarily useful with scope=all.
          schema:
            type: integer
            format: int64
            minimum: 1
        - in: query
          name: diagnosis
          required: false
          description: Optional case-insensitive diagnosis substring filter.
          schema:
            type: string
      responses:
        '200':
          description: Paginated patients list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientListResponse'
              examples:
                ownPatients:
                  summary: Patients from doctor's region
                  value:
                    items:
                      - id: 3
                        patientCode: PT-DEMO-003
                        lastName: Volkov
                        firstName: Andrey
                        middleName: Olegovich
                        birthDate: '1959-11-02'
                        diagnosis: Postoperative observation
                        regionId: 1
                        status: RED
                        valve:
                          name: Abbott Portico
                          size: 27 mm
                          material: Bioprosthetic
                        operationParameters:
                          anesthesia: General anesthesia
                          durationMinutes: 135
                          deliverySystem: Transfemoral
                        medications: Warfarin 2.5 mg daily
                        createdAt: '2026-01-20T14:45:00'
                        lastExaminationAt: '2025-09-01'
                    page: 0
                    limit: 20
                    total: 3
                allPatientsAnonymized:
                  summary: All regions with anonymized names
                  value:
                    items:
                      - id: 5
                        patientCode: PT-DEMO-005
                        lastName: null
                        firstName: null
                        middleName: null
                        birthDate: '1976-12-04'
                        diagnosis: Valve replacement follow-up
                        regionId: 3
                        status: YELLOW
                        valve:
                          name: CoreValve
                          size: 29 mm
                          material: Bioprosthetic
                        operationParameters:
                          anesthesia: General anesthesia
                          durationMinutes: 160
                          deliverySystem: Transaortic
                        medications: Clopidogrel 75 mg daily
                        createdAt: '2026-03-15T16:20:00'
                        lastExaminationAt: '2026-01-05'
                    page: 0
                    limit: 20
                    total: 5
        '401':
          description: Authentication required
        '403':
          description: Only a doctor can list patients
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateValveRequest:
      type: object
      properties:
        name:
          type: string
        size:
          type: string
        material:
          type: string
      required: [name, size, material]
    CreateOperationParametersRequest:
      type: object
      properties:
        anesthesia:
          type: string
        durationMinutes:
          type: integer
        deliverySystem:
          type: string
      required: [anesthesia, durationMinutes, deliverySystem]
    CreatePatientRequest:
      type: object
      properties:
        lastName:
          type: string
        firstName:
          type: string
        middleName:
          type: string
          nullable: true
        birthDate:
          type: string
          format: date
        diagnosis:
          type: string
        regionId:
          type: integer
          format: int64
        valve:
          $ref: '#/components/schemas/CreateValveRequest'
        operationParameters:
          $ref: '#/components/schemas/CreateOperationParametersRequest'
        medications:
          type: string
      required: [lastName, firstName, birthDate, diagnosis, regionId, valve, operationParameters, medications]
...
    ValveResponse:
      type: object
      properties:
        name:
          type: string
        size:
          type: string
        material:
          type: string
      required: [name, size, material]
    OperationParametersResponse:
      type: object
      properties:
        anesthesia:
          type: string
        durationMinutes:
          type: integer
        deliverySystem:
          type: string
      required: [anesthesia, durationMinutes, deliverySystem]
    CreatedPatientResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        patientCode:
          type: string
        temporaryPassword:
          type: string
        lastName:
          type: string
        firstName:
          type: string
        middleName:
          type: string
          nullable: true
        birthDate:
          type: string
          format: date
        diagnosis:
          type: string
        regionId:
          type: integer
          format: int64
        valve:
          $ref: '#/components/schemas/ValveResponse'
        operationParameters:
          $ref: '#/components/schemas/OperationParametersResponse'
        medications:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, patientCode, temporaryPassword, lastName, firstName, birthDate, diagnosis, regionId, valve, operationParameters, medications, createdAt]
    PatientSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        patientCode:
          type: string
        lastName:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        middleName:
          type: string
          nullable: true
        birthDate:
          type: string
          format: date
        diagnosis:
          type: string
        regionId:
          type: integer
          format: int64
        status:
          type: string
          enum: [GREEN, YELLOW, RED]
        valve:
          $ref: '#/components/schemas/ValveResponse'
        operationParameters:
          $ref: '#/components/schemas/OperationParametersResponse'
        medications:
          type: string
        createdAt:
          type: string
          format: date-time
        lastExaminationAt:
          type: string
          format: date
          nullable: true
      required: [id, patientCode, birthDate, diagnosis, regionId, status, valve, operationParameters, medications, createdAt]
    PatientListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PatientSummaryResponse'
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
          format: int64
      required: [items, page, limit, total]
...